% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{create_feedback_tiles}
\alias{create_feedback_tiles}
\title{Create feedback tiles.}
\usage{
create_feedback_tiles(guess, target)
}
\arguments{
\item{guess}{A 5-character string representing the user's guessed word.}

\item{target}{A 5-character string representing the target word.}
}
\value{
A Shiny object containing the feedback tiles.
}
\description{
Generates colored feedback tiles to indicate whether the letters are
correct, present, or wrong.
}
\examples{
create_feedback_tiles("labas", "batas")

}
