% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_zodelis.R
\name{launch_zodelis}
\alias{launch_zodelis}
\title{Launch the Lithuanian Wordle Shiny App}
\usage{
launch_zodelis()
}
\description{
This function launches the Lithuanian version of the Wordle game built in Shiny.
}
