% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{Server logic for Žodelis.}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
None
}
\description{
Processes user guesses, updates tile feedback and shows notifications.
}
\examples{
shinyApp(ui, server)

}
