% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{show_feedback}
\alias{show_feedback}
\title{Show a notification for a correct guess}
\usage{
show_feedback()
}
\value{
A notification "Correct guess!".
}
\description{
Show a notification for a correct guess
}
\examples{
if (interactive()) show_feedback()
}
