% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{start_game}
\alias{start_game}
\title{Print "Welcome to WORDLE. Type in your guess:"}
\usage{
start_game()
}
\value{
No return value.
}
\description{
Print "Welcome to WORDLE. Type in your guess:"
}
\details{
Prints "Welcome to WORDLE. Type in your guess:"
}
\examples{
start_game()

# You can call the function like this:
# start_game()

}
