% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{words_lt}
\alias{words_lt}
\title{Lithuanian 5-letter word list}
\format{
character vector
}
\source{
data/words_lt.rda
}
\usage{
words_lt
}
\description{
A character vector containing valid 5-letter Lithuanian words.
}
\keyword{data}
